shipping-consumer/requirements.txt
kafka-python==2.0.2
flake8==3.9.0
python-dotenv==0.19.0
reportlab==4.0.8
python-barcode==0.15.1